src/music_mcp_server.py
#!/usr/bin/env python3
"""
ElevenLabs Music MCP Server
Exposes music generation tools through the Model Context Protocol

Author: Jeremy Bradford
Date: 2025-11-16
Project: P171_elevenlabs-music-mcp
"""

import asyncio
import json
import sys
import os
import traceback
from typing import Any, Dict, List, Optional
from pathlib import Path
from dotenv import load_dotenv

# MCP imports
from mcp.server import Server
from mcp.server.stdio import stdio_server
from mcp.types import Tool, TextContent

# Local imports (to be implemented)
# from music_generator import MusicGenerator
# from composition_planner import CompositionPlanner  
# from context_analyzer import ContextAnalyzer
# from file_manager import FileManager
# from preference_learner import PreferenceLearner
# from config_manager import Config

# Load environment variables
load_dotenv()


class MusicMCPServer:
    """
    Main MCP server for music generation.
    Exposes tools for AI agents to generate, manage, and learn from music.
    """
    
    def __init__(self):
        """Initialize the MCP server and all components."""
        self.server = Server("elevenlabs-music")
        
        # TODO: Initialize components
        # self.config = Config.from_env()
        # self.generator = MusicGenerator(self.config.api_key)
        # self.planner = CompositionPlanner()
        # self.context_analyzer = ContextAnalyzer()
        # self.file_manager = FileManager(self.config.music_output_dir)
        # self.preference_learner = PreferenceLearner(self.config.preference_storage_path)
        
        # Register all MCP tools
        self.register_tools()
        
    def register_tools(self):
        """Register all available MCP tools."""
        
        # Tool 1: Simple music generation
        @self.server.tool()
        async def generate_music_simple(
            prompt: str,
            duration_ms: int = None,
            output_format: str = "mp3_44100_128",
            metadata: dict = None
        ) -> dict:
            """
            Generate music from a simple text prompt.
            
            Args:
                prompt: Natural language description of desired music
                duration_ms: Optional duration (10000-300000ms)
                output_format: Audio format (default: mp3_44100_128)
                metadata: Optional metadata tags
                
            Returns:
                dict with audio_path, composition_plan, and metadata
            """
            try:
                # TODO: Implement
                return {
                    "success": False,
                    "error": "Not implemented yet"
                }
            except Exception as e:
                return {
                    "success": False,
                    "error": str(e),
                    "traceback": traceback.format_exc()
                }
        
        # Tool 2: Create composition plan
        @self.server.tool()
        async def create_composition_plan(
            prompt: str,
            total_duration_ms: int,
            sections: list = None,
            mood_progression: str = None
        ) -> dict:
            """
            Generate a structured composition plan before creating music.
            
            Args:
                prompt: High-level description
                total_duration_ms: Total length desired
                sections: Optional pre-defined sections
                mood_progression: Optional mood flow (e.g., "calm to energetic")
                
            Returns:
                dict with composition_plan structure
            """
            try:
                # TODO: Implement
                return {
                    "success": False,
                    "error": "Not implemented yet"
                }
            except Exception as e:
                return {
                    "success": False,
                    "error": str(e)
                }
        
        # Tool 3: Generate from composition plan
        @self.server.tool()
        async def generate_music_structured(
            composition_plan: dict,
            strict_duration: bool = False,
            metadata: dict = None
        ) -> dict:
            """
            Generate music from a detailed composition plan.
            
            Args:
                composition_plan: Structured plan from create_composition_plan
                strict_duration: Enforce exact section durations
                metadata: Optional metadata tags
                
            Returns:
                dict with audio_path and generation details
            """
            try:
                # TODO: Implement
                return {
                    "success": False,
                    "error": "Not implemented yet"
                }
            except Exception as e:
                return {
                    "success": False,
                    "error": str(e)
                }
        
        # Tool 4: Analyze mood for music suggestion
        @self.server.tool()
        async def analyze_mood_for_music(
            context: str,
            activity: str = None,
            time_of_day: str = None,
            recent_music: list = None
        ) -> dict:
            """
            Analyze context and suggest appropriate music parameters.
            
            Args:
                context: Text to analyze for mood
                activity: Optional activity type (e.g., "coding", "writing")
                time_of_day: Optional time context (e.g., "morning", "evening")
                recent_music: Optional list of recently generated tracks
                
            Returns:
                dict with suggested_prompt, duration, mood_analysis, reasoning
            """
            try:
                # TODO: Implement
                return {
                    "success": False,
                    "error": "Not implemented yet"
                }
            except Exception as e:
                return {
                    "success": False,
                    "error": str(e)
                }
        
        # Tool 5: Journal entry with music (Phase 3)
        @self.server.tool()
        async def generate_journal_entry_with_music(
            conversation_history: list = None,
            date: str = None,
            mood_override: str = None,
            music_duration_ms: int = 60000
        ) -> dict:
            """
            Create a daily journal entry with matching soundtrack.
            
            Args:
                conversation_history: Recent conversations to analyze
                date: ISO date string (default: today)
                mood_override: Optional forced mood
                music_duration_ms: Length of soundtrack
                
            Returns:
                dict with journal_entry, music details, combined_path
            """
            try:
                # TODO: Implement (Phase 3)
                return {
                    "success": False,
                    "error": "Phase 3 feature - not yet implemented"
                }
            except Exception as e:
                return {
                    "success": False,
                    "error": str(e)
                }
        
        print("✅ Registered 5 MCP tools", file=sys.stderr)
    
    async def run(self):
        """Start the MCP server with stdio transport."""
        async with stdio_server() as (read_stream, write_stream):
            await self.server.run(
                read_stream,
                write_stream,
                self.server.create_initialization_options()
            )


async def main():
    """Main entry point for the MCP server."""
    print("🎵 Starting ElevenLabs Music MCP Server...", file=sys.stderr)
    
    # Check for API key
    api_key = os.getenv("ELEVENLABS_API_KEY")
    if not api_key:
        print("❌ ERROR: ELEVENLABS_API_KEY not set", file=sys.stderr)
        print("   Set it with: export ELEVENLABS_API_KEY='your-key'", file=sys.stderr)
        sys.exit(1)
    
    print(f"✅ API key found", file=sys.stderr)
    
    # Create and run server
    server = MusicMCPServer()
    await server.run()


if __name__ == "__main__":
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print("\n👋 Shutting down gracefully...", file=sys.stderr)
    except Exception as e:
        print(f"\n❌ Fatal error: {e}", file=sys.stderr)
        traceback.print_exc()
        sys.exit(1)
