requirements.txt
# Core MCP Dependencies
mcp>=0.9.0
pydantic>=2.0.0
anthropic>=0.34.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.27.0

# Audio Processing (if needed for future features)
# pydub>=0.25.1

# Data & Storage
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
mypy>=1.7.0
ruff>=0.1.0

# Optional: Sentiment Analysis for Mood Detection
# transformers>=4.35.0
# torch>=2.1.0